requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
pydub==0.25.1
pytaglib==2.0.0
python-slugify==8.0.1
requests==2.31.0
text-unidecode==1.3
urllib3==2.0.3
